@@ -21,7 +21,7 @@ def get_version():
 setup(
     name='ultralytics',  # name of pypi package
     version=get_version(),  # version of pypi package
-    python_requires='>=3.7',
+    python_requires='>=3.8',
     license='AGPL-3.0',
     description=('Ultralytics YOLOv8 for SOTA object detection, multi-object tracking, instance segmentation, '
                  'pose estimation and image classification.'),
@@ -39,6 +39,7 @@ setup(
     install_requires=REQUIREMENTS,
     extras_require={
         'dev': [
+            'ipython',
             'check-manifest',
             'pytest',
             'pytest-cov',
@@ -46,10 +47,10 @@ setup(
             'mkdocs-material',
             'mkdocstrings[python]',
             'mkdocs-redirects',  # for 301 redirects
-            'mkdocs-ultralytics-plugin>=0.0.22',  # for meta descriptions and images, dates and authors
+            'mkdocs-ultralytics-plugin>=0.0.25',  # for meta descriptions and images, dates and authors
         ],
         'export': [
-            'coremltools>=6.0,<=6.2',
+            'coremltools>=7.0.b1',
             'openvino-dev>=2023.0',
             'tensorflowjs',  # automatically installs tensorflow
         ], },
