@@ -1,3 +1,5 @@
+# Ultralytics YOLO 🚀, AGPL-3.0 license
+
 import json
 from collections import defaultdict
 from pathlib import Path
@@ -115,6 +117,97 @@ def convert_coco(labels_dir='../coco/annotations/', use_segments=False, use_keyp
                     file.write(('%g ' * len(line)).rstrip() % line + '\n')
 
 
+def convert_dota_to_yolo_obb(dota_root_path: str):
+    """
+    Converts DOTA dataset annotations to YOLO OBB (Oriented Bounding Box) format.
+
+    The function processes images in the 'train' and 'val' folders of the DOTA dataset. For each image, it reads the
+    associated label from the original labels directory and writes new labels in YOLO OBB format to a new directory.
+
+    Args:
+        dota_root_path (str): The root directory path of the DOTA dataset.
+
+    Example:
+        ```python
+        from ultralytics.data.converter import convert_dota_to_yolo_obb
+
+        convert_dota_to_yolo_obb('path/to/DOTA')
+        ```
+
+    Notes:
+        The directory structure assumed for the DOTA dataset:
+            - DOTA
+                - images
+                    - train
+                    - val
+                - labels
+                    - train_original
+                    - val_original
+
+        After the function execution, the new labels will be saved in:
+            - DOTA
+                - labels
+                    - train
+                    - val
+    """
+    dota_root_path = Path(dota_root_path)
+
+    # Class names to indices mapping
+    class_mapping = {
+        'plane': 0,
+        'ship': 1,
+        'storage-tank': 2,
+        'baseball-diamond': 3,
+        'tennis-court': 4,
+        'basketball-court': 5,
+        'ground-track-field': 6,
+        'harbor': 7,
+        'bridge': 8,
+        'large-vehicle': 9,
+        'small-vehicle': 10,
+        'helicopter': 11,
+        'roundabout': 12,
+        'soccer ball-field': 13,
+        'swimming-pool': 14,
+        'container-crane': 15,
+        'airport': 16,
+        'helipad': 17}
+
+    def convert_label(image_name, image_width, image_height, orig_label_dir, save_dir):
+        orig_label_path = orig_label_dir / f'{image_name}.txt'
+        save_path = save_dir / f'{image_name}.txt'
+
+        with orig_label_path.open('r') as f, save_path.open('w') as g:
+            lines = f.readlines()
+            for line in lines:
+                parts = line.strip().split()
+                if len(parts) < 9:
+                    continue
+                class_name = parts[8]
+                class_idx = class_mapping[class_name]
+                coords = [float(p) for p in parts[:8]]
+                normalized_coords = [
+                    coords[i] / image_width if i % 2 == 0 else coords[i] / image_height for i in range(8)]
+                formatted_coords = ['{:.6g}'.format(coord) for coord in normalized_coords]
+                g.write(f"{class_idx} {' '.join(formatted_coords)}\n")
+
+    for phase in ['train', 'val']:
+        image_dir = dota_root_path / 'images' / phase
+        orig_label_dir = dota_root_path / 'labels' / f'{phase}_original'
+        save_dir = dota_root_path / 'labels' / phase
+
+        save_dir.mkdir(parents=True, exist_ok=True)
+
+        image_paths = list(image_dir.iterdir())
+        for image_path in tqdm(image_paths, desc=f'Processing {phase} images'):
+            if image_path.suffix != '.png':
+                continue
+            image_name_without_ext = image_path.stem
+            img = cv2.imread(str(image_path))
+            h, w = img.shape[:2]
+            convert_label(image_name_without_ext, w, h, orig_label_dir, save_dir)
+
+
 def rle2polygon(segmentation):
     """
     Convert Run-Length Encoding (RLE) mask to polygon coordinates.
@@ -207,24 +300,3 @@ def merge_multi_segment(segments):
                     nidx = abs(idx[1] - idx[0])
                     s.append(segments[i][nidx:])
     return s
-
-
-def delete_dsstore(path='../datasets'):
-    """Delete Apple .DS_Store files in the specified directory and its subdirectories."""
-    from pathlib import Path
-
-    files = list(Path(path).rglob('.DS_store'))
-    print(files)
-    for f in files:
-        f.unlink()
-
-
-if __name__ == '__main__':
-    source = 'COCO'
-
-    if source == 'COCO':
-        convert_coco(
-            '../datasets/coco/annotations',  # directory with *.json
-            use_segments=False,
-            use_keypoints=True,
-            cls91to80=False)
