@@ -12,18 +12,20 @@ class FastSAM(Model):
     """
     FastSAM model interface.
 
-    Usage - Predict:
+    Example:
+        ```python
         from ultralytics import FastSAM
 
         model = FastSAM('last.pt')
         results = model.predict('ultralytics/assets/bus.jpg')
+        ```
     """
 
     def __init__(self, model='FastSAM-x.pt'):
         """Call the __init__ method of the parent class (YOLO) with the updated default model"""
-        if model == 'FastSAM.pt':
+        if str(model) == 'FastSAM.pt':
             model = 'FastSAM-x.pt'
-        assert Path(model).suffix != '.yaml', 'FastSAM models only support pre-trained models.'
+        assert Path(model).suffix not in ('.yaml', '.yml'), 'FastSAM models only support pre-trained models.'
         super().__init__(model=model, task='segment')
 
     @property
