@@ -14,7 +14,7 @@ TorchScript             | `torchscript`             | yolov8n.torchscript
 ONNX                    | `onnx`                    | yolov8n.onnx
 OpenVINO                | `openvino`                | yolov8n_openvino_model/
 TensorRT                | `engine`                  | yolov8n.engine
-CoreML                  | `coreml`                  | yolov8n.mlmodel
+CoreML                  | `coreml`                  | yolov8n.mlpackage
 TensorFlow SavedModel   | `saved_model`             | yolov8n_saved_model/
 TensorFlow GraphDef     | `pb`                      | yolov8n.pb
 TensorFlow Lite         | `tflite`                  | yolov8n.tflite
@@ -192,7 +192,7 @@ class ProfileModels:
         output = []
         for file in files:
             engine_file = file.with_suffix('.engine')
-            if file.suffix in ('.pt', '.yaml'):
+            if file.suffix in ('.pt', '.yaml', '.yml'):
                 model = YOLO(str(file))
                 model.fuse()  # to report correct params and GFLOPs in model.info()
                 model_info = model.info()
@@ -229,7 +229,7 @@ class ProfileModels:
             if path.is_dir():
                 extensions = ['*.pt', '*.onnx', '*.yaml']
                 files.extend([file for ext in extensions for file in glob.glob(str(path / ext))])
-            elif path.suffix in {'.pt', '.yaml'}:  # add non-existing
+            elif path.suffix in ('.pt', '.yaml', '.yml'):  # add non-existing
                 files.append(str(path))
             else:
                 files.extend(glob.glob(str(path)))
